requirements.txt
Django==2.1.1
Jinja2==2.10
django-filter==2.0.0
djangorestframework==3.8.2
